Change /inbox

+++ ik/captcha/image.go
@@ -62,7 +62,7 @@ func NewImage(id string, digits []byte, width, height int) *Image {
 	// Draw strike-through line.
 	m.strikeThrough()
 	// Apply wave distortion.
-	m.distort(m.rng.Float(5, 10), m.rng.Float(100, 200))
+	m.distort(m.rng.Float(3, 8), m.rng.Float(100, 200))
 	// Fill image with random circles.
 	m.fillWithCircles(circleCount, m.dotSize)
 	return m
